Rename module pandastable to dataframe
